Moving files to new structure
